[UPDATE] add ConfirmationViewModel
