+++ app/src/main/java/com/immobylette/appmobile/MainActivity.kt
@@ -10,6 +10,7 @@ import com.immobylette.appmobile.agent.current.CurrentAgentViewModel
 import com.immobylette.appmobile.agent.selection.AgentSelectionViewModel
 import com.immobylette.appmobile.agent.selection.agentSelectionNavigation
 import com.immobylette.appmobile.agent.selection.navigateToAgentSelection
+import com.immobylette.appmobile.confirmation.ConfirmationViewModel
 import com.immobylette.appmobile.confirmation.confirmationNavigation
 import com.immobylette.appmobile.confirmation.navigateToConfirmation
 import com.immobylette.appmobile.loading.loadingNavigation
@@ -35,6 +36,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             val agentSelectionViewModel by viewModels<AgentSelectionViewModel>()
             val propertySelectionViewModel by viewModels<PropertySelectionViewModel>()
+            val confirmationViewModel by viewModels<ConfirmationViewModel>()
             val navController = rememberNavController()
 
             ImmobyletteappmobileTheme {
@@ -60,6 +62,7 @@ class MainActivity : ComponentActivity() {
                     )
 
                     confirmationNavigation(
+                        confirmationViewModel = confirmationViewModel,
                         currentPropertyViewModel = currentPropertyViewModel,
                         currentInventoryViewModel = currentInventoryViewModel,
                         currentAgentViewModel = currentAgentViewModel,
